// extension/src/components/store/data-types.ts
export interface IPerson {
  readonly id: string;
  readonly name: string;
  readonly emailAddresses: string[];
  readonly imageUrl?: string;
  readonly notes?: string;
  readonly googleIds: string[];
  readonly isCurrentUser: number; // needs to be a number to be a valid index
  readonly isInContacts: boolean;
  readonly etag?: string;
  readonly dateAdded: Date;
}